Make display customized + add mobile compatibility - JT
Removed bloat like bootstrap :)
so we can handle styles ourselves.
Removed JQuery as we are not using it.
CSS was added to display both mobile and desktop
page menus + to make displays custom.
Using CSS3 features for dynamic rendering + display of the navbar. i.e. there is no JS for nav.